feat(uikit, venue-details, app): 🔥 Implement venue details screen

+++ api/src/main/java/ru/stakancheck/api/models/CurrentWeatherDTO.kt
@@ -100,8 +100,8 @@ data class Main(
  */
 @Serializable
 data class Wind(
-    val speed: Int,
-    val deg: Int
+    val speed: Double,
+    val deg: Double
 )
 
 /**
